+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 - [BDN-626](https://appodeal.atlassian.net/browse/BDN-626) Integrated VkAds (myTarget) as Ad Adapter
 - [BDN-483](https://appodeal.atlassian.net/browse/BDN-483) Integrated CPM mode for Vungle adapter;
 - [BDN-716](https://appodeal.atlassian.net/browse/BDN-716) Integrated CPM mode for BigoAds adapter;
-- [BDN-719](https://appodeal.atlassian.net/browse/BDN-719) Integrated Yandex as Ad Adapter
+- [BDN-719](https://appodeal.atlassian.net/browse/BDN-719) Integrated Yandex (cpm) as Ad Adapter
+- [BDN-482](https://appodeal.atlassian.net/browse/BDN-482) Integrated IronSource (cpm) as Ad Adapter
 
 # Next 0.7.0-next.1
 ## Features: